requirements.txt
tweepy==4.14.0
pandas==2.0.3
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
wheel==0.41.2
setuptools==68.0.0 